refactor: extract WFS 2.0 operation sub-handlers from Wfs20Handler (#1144)
Code quality audit flagged Wfs20Handler.cs at 2,751 LOC. The handler is
already split into partials (Capabilities, StoredQueries, Transaction,
Legacy) yet the main file still carried the GetFeature and GetPropertyValue
operations plus the bulk of the shared response-building infrastructure.

This change moves the GetFeature and GetPropertyValue operations and their
operation-exclusive helpers into focused partial files:

- Wfs20Handler.GetFeature.cs (509 LOC): HandleGetFeatureAsync,
  HandleGetFeatureXmlQueriesAsync, BuildXmlQueryPlansAsync,
  TryGetMultiQueryXmlRequest.
- Wfs20Handler.GetPropertyValue.cs (484 LOC): HandleGetPropertyValueAsync,
  BuildLayerValuePlansAsync, BuildValueQueryAsync,
  BuildValueCollectionResultAsync, BuildValueJsonResultAsync,
  ResolveValueReference, ExtractValue, ConvertValueReferenceToInvariantString,
  IsSupportedValueOutputFormat, CreateEmptyValueCollectionResult,
  CreateEmptyValueCollectionXmlResult, and the LayerValuePlan,
  LayerValuePlanSet, ValueReferenceResolution nested types.

Wfs20Handler.cs is reduced from 2,751 LOC to 1,811 LOC (-940, -34%).
Public surface is unchanged (HandleGetFeatureAsync /
HandleGetPropertyValueAsync remain public on Wfs20Handler) so callers and
DI registration in Wfs20ServiceCollectionExtensions need no edits.

Note on the audit's "sealed sub-handler class" guidance: the GetFeature
and GetPropertyValue paths depend on ~16 shared private helpers
(BuildFeatureQueryAsync, BuildLayerQueryPlansAsync,
GetPublishedFeatureTypesAsync, ResolveRequestedFeatureTypes,
BuildFeatureCollectionResponseMetadata, BuildGmlResultAsync,
BuildJsonResultAsync, BuildCsvResultAsync,
BuildPagedGetFeatureResultAsync, ShouldUsePagedGetFeatureFastPath,
CreateEmptyFeatureCollectionResult, CreateHitsFeatureCollectionResult,
NormalizeFilterInputs, TranslateFesFilter, ParseBboxFilter,
ResolveRequestedOutputSridAsync, ...) and ~6 nested private types
(WfsFeatureTypeDescriptor, LayerQueryPlan, LayerQueryPlanSet,
PagingLinks, FeatureCollectionResponseMetadata, PagedGetFeatureResult)
that are also consumed by the Legacy, StoredQueries, Capabilities, and
Transaction partials. A full sealed-class extraction would have
required exposing all of those as `internal` accessors or duplicating
them, both of which carry meaningful regression risk against the
168/168 CITE compliance baseline. Following the audit guidance to
prefer a clean partial extraction over a broken full extraction, this
PR delivers the partial-file split as a low-risk interim that already
addresses the main-file size finding. A follow-up can lift the shared
helpers into a stateless service if a true sealed split is desired.

Baseline tests (before): Failed: 1, Passed: 117, Total: 118.
After refactor:           Failed: 1, Passed: 117, Total: 118.
The single failing test (Wfs_GetCapabilities_WithProjectedExtent_
UsesTransformFallbackForWgs84BoundingBox) is pre-existing and
unrelated to this change.
