+++ src/Honua.Server/Features/Protocols/Ogc/Classic/Wfs20/Services/Wfs20Handler.cs
@@ -106,424 +106,11 @@ internal sealed partial class Wfs20Handler
             "id");
     }
 
-    public async Task<IResult> HandleGetFeatureAsync(
-        HttpContext context,
-        string? typeNames,
-        string? outputFormat,
-        string? count,
-        string? startIndex,
-        string? sortBy,
-        string? bbox,
-        string? filter,
-        string? resourceId,
-        string? propertyName,
-        string? srsName,
-        string? resultType,
-        CancellationToken cancellationToken = default)
-    {
-        using var activity = HonuaTelemetry.ActivitySource.StartActivity(
-            "wfs20.get_feature", ActivityKind.Internal);
-        activity?.SetTag(HonuaTelemetry.Tags.Protocol, HonuaTelemetry.Protocols.Wfs20);
-        activity?.SetTag(HonuaTelemetry.Tags.Operation, Wfs20Utilities.Operations.GetFeature);
-        activity?.SetTag("wfs.type_names", typeNames ?? "ALL");
-        activity?.SetTag("wfs.output_format", outputFormat ?? Wfs20Utilities.OutputFormats.Default);
-
-        var normalizedFormat = Wfs20Utilities.OutputFormats.NormalizeOutputFormat(outputFormat);
-        if (!Wfs20Utilities.TryNormalizeResultType(resultType, out var normalizedResultType))
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                $"Invalid RESULTTYPE parameter '{resultType}'. Supported values are 'results' and 'hits'.",
-                "resultType");
-        }
-
-        if (!Wfs20Utilities.TryParseCount(count, out var maxFeatures))
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                $"Invalid COUNT parameter '{count}'. COUNT must be a non-negative integer.",
-                "count");
-        }
-
-        if (!Wfs20Utilities.TryParseStartIndex(startIndex, out var offset))
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                $"Invalid STARTINDEX parameter '{startIndex}'. STARTINDEX must be a non-negative integer.",
-                "startIndex");
-        }
-
-        var requestedTypes = Wfs20Utilities.ParseTypeNames(typeNames);
-        var isHitsRequest = string.Equals(normalizedResultType, "hits", StringComparison.OrdinalIgnoreCase);
-
-        Wfs20Log.GetFeatureRequested(_logger, typeNames ?? "ALL", normalizedFormat);
-
-        try
-        {
-            if (!IsSupportedFeatureOutputFormat(normalizedFormat))
-            {
-                Wfs20Log.UnsupportedOutputFormatRequested(_logger, outputFormat ?? normalizedFormat);
-                return Wfs20ErrorResults.CreateBadRequest(
-                    context,
-                    "InvalidParameterValue",
-                    $"Unsupported output format '{outputFormat}'. Supported formats: {string.Join(", ", Wfs20Utilities.OutputFormats.All)}",
-                    "outputFormat");
-            }
-
-            if (TryGetMultiQueryXmlRequest(context, out var xmlQueries))
-            {
-                return await HandleGetFeatureXmlQueriesAsync(
-                    context,
-                    xmlQueries,
-                    normalizedFormat,
-                    outputFormat,
-                    maxFeatures,
-                    offset,
-                    sortBy,
-                    bbox,
-                    resourceId,
-                    srsName,
-                    normalizedResultType,
-                    isHitsRequest,
-                    cancellationToken);
-            }
-
-            var publishedTypes = await GetPublishedFeatureTypesAsync(context, cancellationToken);
-            var unknownTypes = GetUnknownRequestedFeatureTypes(publishedTypes, requestedTypes);
-            if (unknownTypes.Length > 0)
-            {
-                var requestedTypeMessage = unknownTypes.Length == 1
-                    ? $"Unknown feature type '{unknownTypes[0]}'."
-                    : $"Unknown feature types: {string.Join(", ", unknownTypes.Select(type => $"'{type}'"))}.";
-                return Wfs20ErrorResults.CreateBadRequest(
-                    context,
-                    "InvalidParameterValue",
-                    requestedTypeMessage,
-                    "typeNames");
-            }
-
-            var selectedTypes = ResolveRequestedFeatureTypes(publishedTypes, requestedTypes);
-            var wfsUrl = $"{BaseUrlResolver.GetBaseUrl(context)}/wfs";
-            if (selectedTypes.Length == 0)
-            {
-                return isHitsRequest
-                    ? CreateHitsFeatureCollectionResult(0)
-                    : CreateEmptyFeatureCollectionResult(normalizedFormat);
-            }
-
-            if (ShouldUsePagedGetFeatureFastPath(selectedTypes, normalizedFormat, isHitsRequest, maxFeatures))
-            {
-                var descriptor = selectedTypes[0];
-                var query = (await BuildFeatureQueryAsync(
-                    descriptor.Layer,
-                    propertyName,
-                    sortBy,
-                    bbox,
-                    filter,
-                    resourceId,
-                    srsName,
-                    enforceResourceIdTypeMatch: true,
-                    requireResourceIdQualifier: selectedTypes.Length > 1,
-                    cancellationToken: cancellationToken).ConfigureAwait(false)) with
-                {
-                    Offset = offset,
-                    Limit = maxFeatures
-                };
-
-                var pagedResult = await BuildPagedGetFeatureResultAsync(
-                    descriptor,
-                    query,
-                    normalizedFormat,
-                    cancellationToken);
-
-                Wfs20Log.GetFeatureReturned(
-                    _logger,
-                    pagedResult.ReturnedCount,
-                    pagedResult.NumberMatchedSummary);
-
-                return pagedResult.Result;
-            }
-
-            var planSet = await BuildLayerQueryPlansAsync(
-                selectedTypes,
-                propertyName,
-                sortBy,
-                bbox,
-                filter,
-                resourceId,
-                srsName,
-                offset,
-                maxFeatures,
-                cancellationToken);
-
-            if (planSet.TotalMatched == 0)
-            {
-                var emptyMetadata = BuildFeatureCollectionResponseMetadata(
-                    wfsUrl,
-                    selectedTypes,
-                    outputFormat,
-                    count,
-                    sortBy,
-                    bbox,
-                    filter,
-                    resourceId,
-                    propertyName,
-                    srsName,
-                    normalizedResultType,
-                    offset,
-                    maxFeatures,
-                    0,
-                    0);
-                return isHitsRequest
-                    ? CreateHitsFeatureCollectionResult(0, emptyMetadata.SchemaLocation, emptyMetadata.PagingLinks.Next, emptyMetadata.PagingLinks.Previous)
-                    : CreateEmptyFeatureCollectionResult(normalizedFormat, emptyMetadata.SchemaLocation, emptyMetadata.PagingLinks.Next, emptyMetadata.PagingLinks.Previous);
-            }
-
-            if (isHitsRequest)
-            {
-                var hitsMetadata = BuildFeatureCollectionResponseMetadata(
-                    wfsUrl,
-                    selectedTypes,
-                    outputFormat,
-                    count,
-                    sortBy,
-                    bbox,
-                    filter,
-                    resourceId,
-                    propertyName,
-                    srsName,
-                    normalizedResultType,
-                    offset,
-                    maxFeatures,
-                    planSet.TotalMatched,
-                    0);
-                if (_logger.IsEnabled(LogLevel.Information))
-                {
-                    var matchedSummary = planSet.TotalMatched.ToString(CultureInfo.InvariantCulture);
-                    Wfs20Log.GetFeatureReturned(_logger, 0, matchedSummary);
-                }
-                return CreateHitsFeatureCollectionResult(
-                    planSet.TotalMatched,
-                    hitsMetadata.SchemaLocation,
-                    hitsMetadata.PagingLinks.Next,
-                    hitsMetadata.PagingLinks.Previous);
-            }
-
-            var expectedReturnedCount = planSet.Plans.Sum(plan => plan.Query.Limit ?? 0);
-            var responseMetadata = BuildFeatureCollectionResponseMetadata(
-                wfsUrl,
-                selectedTypes,
-                outputFormat,
-                count,
-                sortBy,
-                bbox,
-                filter,
-                resourceId,
-                propertyName,
-                srsName,
-                normalizedResultType,
-                offset,
-                maxFeatures,
-                planSet.TotalMatched,
-                expectedReturnedCount);
-
-            var (result, returnedCount) = normalizedFormat switch
-            {
-                Wfs20Utilities.OutputFormats.Csv => await BuildCsvResultAsync(planSet, cancellationToken),
-                Wfs20Utilities.OutputFormats.GeoJson or Wfs20Utilities.OutputFormats.Json => await BuildJsonResultAsync(planSet, normalizedFormat, cancellationToken),
-                _ => await BuildGmlResultAsync(
-                    planSet,
-                    responseMetadata.SchemaLocation,
-                    responseMetadata.PagingLinks.Next,
-                    responseMetadata.PagingLinks.Previous,
-                    cancellationToken)
-            };
-
-            if (_logger.IsEnabled(LogLevel.Information))
-            {
-                var matchedSummary = planSet.TotalMatched.ToString(CultureInfo.InvariantCulture);
-                Wfs20Log.GetFeatureReturned(
-                    _logger,
-                    returnedCount,
-                    matchedSummary);
-            }
-
-            return result;
-        }
-        catch (WfsQueryException ex)
-        {
-            Wfs20Log.ParameterValidationFailed(_logger, ex.Message);
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                ex.ExceptionCode,
-                ex.Message,
-                ex.Locator);
-        }
-        catch (Exception ex) when (ex is ArgumentException or NotSupportedException or Fes20ParseException)
-        {
-            Wfs20Log.ParameterValidationFailed(_logger, ex.Message);
-            var exceptionCode = ex.Message.Contains("boundedBy", StringComparison.OrdinalIgnoreCase)
-                ? "OperationProcessingFailed"
-                : "InvalidParameterValue";
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                exceptionCode,
-                "Invalid WFS parameter value; see logs for details.");
-        }
-        catch (Exception ex)
-        {
-            Wfs20Log.DatabaseQueryFailed(_logger, Wfs20Utilities.Operations.GetFeature, ex.Message);
-            return StandardErrorHelpers.CreateInternalServerError(context, "Failed to process GetFeature request.");
-        }
-    }
-
-    public async Task<IResult> HandleGetPropertyValueAsync(
-        HttpContext context,
-        string? typeNames,
-        string? valueReference,
-        bool valueReferenceSpecified,
-        string? outputFormat,
-        string? count,
-        string? startIndex,
-        string? bbox,
-        string? filter,
-        string? resourceId,
-        string? srsName,
-        CancellationToken cancellationToken = default)
-    {
-        using var activity = HonuaTelemetry.ActivitySource.StartActivity(
-            "wfs20.get_property_value", ActivityKind.Internal);
-        activity?.SetTag(HonuaTelemetry.Tags.Protocol, HonuaTelemetry.Protocols.Wfs20);
-        activity?.SetTag(HonuaTelemetry.Tags.Operation, Wfs20Utilities.Operations.GetPropertyValue);
-        activity?.SetTag("wfs.type_names", typeNames ?? "ALL");
-        activity?.SetTag("wfs.value_reference", valueReference ?? "unknown");
-
-        if (!valueReferenceSpecified)
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "MissingParameterValue",
-                "Missing required 'valueReference' parameter for GetPropertyValue.",
-                "valueReference");
-        }
-
-        if (string.IsNullOrWhiteSpace(valueReference))
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                "The 'valueReference' parameter must not be empty for GetPropertyValue.",
-                "valueReference");
-        }
-
-        var normalizedOutputFormat = Wfs20Utilities.OutputFormats.NormalizeOutputFormat(outputFormat);
-        if (!IsSupportedValueOutputFormat(normalizedOutputFormat))
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                $"Unsupported output format '{outputFormat}'. GetPropertyValue supports application/gml+xml, application/json, and application/geo+json.",
-                "outputFormat");
-        }
-
-        var requestedTypes = Wfs20Utilities.ParseTypeNames(typeNames);
-        if (!Wfs20Utilities.TryParseCount(count, out var maxFeatures))
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                $"Invalid COUNT parameter '{count}'. COUNT must be a non-negative integer.",
-                "count");
-        }
-
-        if (!Wfs20Utilities.TryParseStartIndex(startIndex, out var offset))
-        {
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                $"Invalid STARTINDEX parameter '{startIndex}'. STARTINDEX must be a non-negative integer.",
-                "startIndex");
-        }
-
-        Wfs20Log.GetPropertyValueRequested(_logger, valueReference, typeNames ?? "ALL");
-
-        try
-        {
-            var publishedTypes = await GetPublishedFeatureTypesAsync(context, cancellationToken);
-            var unknownTypes = GetUnknownRequestedFeatureTypes(publishedTypes, requestedTypes);
-            if (unknownTypes.Length > 0)
-            {
-                var requestedTypeMessage = unknownTypes.Length == 1
-                    ? $"Unknown feature type '{unknownTypes[0]}'."
-                    : $"Unknown feature types: {string.Join(", ", unknownTypes.Select(type => $"'{type}'"))}.";
-                return Wfs20ErrorResults.CreateBadRequest(
-                    context,
-                    "InvalidParameterValue",
-                    requestedTypeMessage,
-                    "typeNames");
-            }
-
-            var selectedTypes = ResolveRequestedFeatureTypes(publishedTypes, requestedTypes);
-            if (selectedTypes.Length == 0)
-            {
-                return CreateEmptyValueCollectionResult(normalizedOutputFormat);
-            }
-
-            var planSet = await BuildLayerValuePlansAsync(
-                selectedTypes,
-                valueReference,
-                bbox,
-                filter,
-                resourceId,
-                srsName,
-                offset,
-                maxFeatures,
-                cancellationToken);
-
-            if (planSet.TotalMatched == 0)
-            {
-                return CreateEmptyValueCollectionResult(normalizedOutputFormat);
-            }
-
-            var (result, returnedCount) = string.Equals(normalizedOutputFormat, Wfs20Utilities.OutputFormats.Gml32, StringComparison.OrdinalIgnoreCase)
-                ? await BuildValueCollectionResultAsync(planSet, cancellationToken)
-                : await BuildValueJsonResultAsync(planSet, normalizedOutputFormat, cancellationToken);
-            Wfs20Log.GetPropertyValueReturned(_logger, returnedCount);
-            return result;
-        }
-        catch (WfsQueryException ex)
-        {
-            Wfs20Log.ParameterValidationFailed(_logger, ex.Message);
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                ex.ExceptionCode,
-                ex.Message,
-                ex.Locator);
-        }
-        catch (Exception ex) when (ex is ArgumentException or NotSupportedException or Fes20ParseException)
-        {
-            Wfs20Log.ParameterValidationFailed(_logger, ex.Message);
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                "Invalid WFS parameter value; see logs for details.");
-        }
-        catch (Exception ex)
-        {
-            Wfs20Log.DatabaseQueryFailed(_logger, Wfs20Utilities.Operations.GetPropertyValue, ex.Message);
-            return StandardErrorHelpers.CreateInternalServerError(context, "Failed to process GetPropertyValue request.");
-        }
-    }
-
     /// <summary>
     /// Processes WFS 2.0 Transaction requests by translating XML actions into the shared edit pipeline.
     /// Behavior reference: ../Honua.Server/src/domain/ogc/wfs/WfsTransactionHandlers.cs
     /// Preserves request order and carries insert handles into InsertResults.
     /// </summary>
-
     private async Task<LayerQueryPlanSet> BuildLayerQueryPlansAsync(
         IReadOnlyList<WfsFeatureTypeDescriptor> featureTypes,
         string? propertyName,
@@ -595,291 +182,6 @@ internal sealed partial class Wfs20Handler
         return new LayerQueryPlanSet(plans.ToImmutable(), totalMatched);
     }
 
-    private async Task<IResult> HandleGetFeatureXmlQueriesAsync(
-        HttpContext context,
-        IReadOnlyList<Wfs20XmlQueryParameters> xmlQueries,
-        string normalizedFormat,
-        string? outputFormat,
-        int maxFeatures,
-        int offset,
-        string? sortBy,
-        string? bbox,
-        string? resourceId,
-        string? fallbackSrsName,
-        string? normalizedResultType,
-        bool isHitsRequest,
-        CancellationToken cancellationToken)
-    {
-        var publishedTypes = await GetPublishedFeatureTypesAsync(context, cancellationToken);
-        var unknownTypes = xmlQueries
-            .SelectMany(query => Wfs20Utilities.ParseTypeNames(query.TypeNames))
-            .Where(requestedType => GetUnknownRequestedFeatureTypes(publishedTypes, new[] { requestedType }).Length > 0)
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-        if (unknownTypes.Length > 0)
-        {
-            var requestedTypeMessage = unknownTypes.Length == 1
-                ? $"Unknown feature type '{unknownTypes[0]}'."
-                : $"Unknown feature types: {string.Join(", ", unknownTypes.Select(type => $"'{type}'"))}.";
-            return Wfs20ErrorResults.CreateBadRequest(
-                context,
-                "InvalidParameterValue",
-                requestedTypeMessage,
-                "typeNames");
-        }
-
-        var planSet = await BuildXmlQueryPlansAsync(
-            publishedTypes,
-            xmlQueries,
-            sortBy,
-            bbox,
-            resourceId,
-            fallbackSrsName,
-            offset,
-            maxFeatures,
-            cancellationToken);
-
-        var selectedTypes = xmlQueries
-            .SelectMany(query => ResolveRequestedFeatureTypes(publishedTypes, Wfs20Utilities.ParseTypeNames(query.TypeNames)))
-            .DistinctBy(static descriptor => descriptor.QualifiedName, StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-        var wfsUrl = $"{BaseUrlResolver.GetBaseUrl(context)}/wfs";
-
-        if (planSet.TotalMatched == 0)
-        {
-            var emptyMetadata = BuildFeatureCollectionResponseMetadata(
-                wfsUrl,
-                selectedTypes,
-                outputFormat,
-                maxFeatures.ToString(CultureInfo.InvariantCulture),
-                sortBy,
-                bbox,
-                null,
-                resourceId,
-                null,
-                fallbackSrsName,
-                normalizedResultType,
-                offset,
-                maxFeatures,
-                0,
-                0);
-            return isHitsRequest
-                ? CreateHitsFeatureCollectionResult(0, emptyMetadata.SchemaLocation, emptyMetadata.PagingLinks.Next, emptyMetadata.PagingLinks.Previous)
-                : CreateEmptyFeatureCollectionResult(normalizedFormat, emptyMetadata.SchemaLocation, emptyMetadata.PagingLinks.Next, emptyMetadata.PagingLinks.Previous);
-        }
-
-        var responseMetadata = BuildFeatureCollectionResponseMetadata(
-            wfsUrl,
-            selectedTypes,
-            outputFormat,
-            maxFeatures.ToString(CultureInfo.InvariantCulture),
-            sortBy,
-            bbox,
-            null,
-            resourceId,
-            null,
-            fallbackSrsName,
-            normalizedResultType,
-            offset,
-            maxFeatures,
-            planSet.TotalMatched,
-            isHitsRequest ? 0 : planSet.Plans.Sum(plan => plan.Query.Limit ?? 0));
-
-        if (isHitsRequest)
-        {
-            return CreateHitsFeatureCollectionResult(
-                planSet.TotalMatched,
-                responseMetadata.SchemaLocation,
-                responseMetadata.PagingLinks.Next,
-                responseMetadata.PagingLinks.Previous);
-        }
-
-        var (result, returnedCount) = normalizedFormat switch
-        {
-            Wfs20Utilities.OutputFormats.Csv => await BuildCsvResultAsync(planSet, cancellationToken),
-            Wfs20Utilities.OutputFormats.GeoJson or Wfs20Utilities.OutputFormats.Json => await BuildJsonResultAsync(planSet, normalizedFormat, cancellationToken),
-            _ => await BuildGmlResultAsync(
-                planSet,
-                responseMetadata.SchemaLocation,
-                responseMetadata.PagingLinks.Next,
-                responseMetadata.PagingLinks.Previous,
-                cancellationToken)
-        };
-
-        if (_logger.IsEnabled(LogLevel.Information))
-        {
-            var matchedSummary = planSet.TotalMatched.ToString(CultureInfo.InvariantCulture);
-            Wfs20Log.GetFeatureReturned(
-                _logger,
-                returnedCount,
-                matchedSummary);
-        }
-
-        return result;
-    }
-
-    private async Task<LayerQueryPlanSet> BuildXmlQueryPlansAsync(
-        ImmutableArray<WfsFeatureTypeDescriptor> publishedTypes,
-        IReadOnlyList<Wfs20XmlQueryParameters> xmlQueries,
-        string? sortBy,
-        string? bbox,
-        string? resourceId,
-        string? fallbackSrsName,
-        int offset,
-        int count,
-        CancellationToken cancellationToken)
-    {
-        var plans = ImmutableArray.CreateBuilder<LayerQueryPlan>();
-        var remainingOffset = offset;
-        var remainingCount = count;
-        long totalMatched = 0;
-
-        foreach (var xmlQuery in xmlQueries)
-        {
-            var selectedTypes = ResolveRequestedFeatureTypes(
-                publishedTypes,
-                Wfs20Utilities.ParseTypeNames(xmlQuery.TypeNames));
-            foreach (var featureType in selectedTypes)
-            {
-                var query = await BuildFeatureQueryAsync(
-                    featureType.Layer,
-                    xmlQuery.PropertyName,
-                    xmlQuery.SortBy ?? sortBy,
-                    bbox,
-                    xmlQuery.Filter,
-                    resourceId,
-                    xmlQuery.SrsName ?? fallbackSrsName,
-                    enforceResourceIdTypeMatch: true,
-                    requireResourceIdQualifier: selectedTypes.Length > 1,
-                    cancellationToken: cancellationToken);
-
-                var layerMatched = await _featureReader.CountAsync(featureType.Layer.Id, query, cancellationToken);
-                totalMatched += layerMatched;
-
-                if (layerMatched == 0)
-                {
-                    continue;
-                }
-
-                if (remainingOffset >= layerMatched)
-                {
-                    remainingOffset -= (int)Math.Min(remainingOffset, layerMatched);
-                    continue;
-                }
-
-                if (remainingCount <= 0)
-                {
-                    continue;
-                }
-
-                var layerOffset = remainingOffset;
-                remainingOffset = 0;
-                var availableCount = layerMatched - layerOffset;
-                var layerLimit = (int)Math.Min(remainingCount, Math.Min(availableCount, int.MaxValue));
-
-                if (layerLimit <= 0)
-                {
-                    continue;
-                }
-
-                plans.Add(new LayerQueryPlan(
-                    featureType,
-                    query with { Offset = layerOffset, Limit = layerLimit },
-                    layerMatched));
-
-                remainingCount -= layerLimit;
-            }
-        }
-
-        return new LayerQueryPlanSet(plans.ToImmutable(), totalMatched);
-    }
-
-    private static bool TryGetMultiQueryXmlRequest(
-        HttpContext context,
-        out Wfs20XmlQueryParameters[] queries)
-    {
-        if (context.Items.TryGetValue(Wfs20DispatcherEndpoint.ParsedXmlQueriesItemKey, out var value) &&
-            value is Wfs20XmlQueryParameters[] parsedQueries &&
-            parsedQueries.Length > 1)
-        {
-            queries = parsedQueries;
-            return true;
-        }
-
-        queries = [];
-        return false;
-    }
-
-    private async Task<LayerValuePlanSet> BuildLayerValuePlansAsync(
-        IReadOnlyList<WfsFeatureTypeDescriptor> featureTypes,
-        string valueReference,
-        string? bbox,
-        string? filter,
-        string? resourceId,
-        string? srsName,
-        int offset,
-        int count,
-        CancellationToken cancellationToken)
-    {
-        var plans = ImmutableArray.CreateBuilder<LayerValuePlan>(featureTypes.Count);
-        var remainingOffset = offset;
-        var remainingCount = count;
-        long totalMatched = 0;
-
-        foreach (var featureType in featureTypes)
-        {
-            var resolvedValueReference = ResolveValueReference(featureType.Layer, valueReference);
-            var query = await BuildValueQueryAsync(
-                featureType.Layer,
-                resolvedValueReference,
-                bbox,
-                filter,
-                resourceId,
-                srsName,
-                enforceResourceIdTypeMatch: true,
-                requireResourceIdQualifier: featureTypes.Count > 1,
-                cancellationToken: cancellationToken);
-
-            var layerMatched = await _featureReader.CountAsync(featureType.Layer.Id, query, cancellationToken);
-            totalMatched += layerMatched;
-
-            if (layerMatched == 0)
-            {
-                continue;
-            }
-
-            if (remainingOffset >= layerMatched)
-            {
-                remainingOffset -= (int)Math.Min(remainingOffset, layerMatched);
-                continue;
-            }
-
-            if (remainingCount <= 0)
-            {
-                continue;
-            }
-
-            var layerOffset = remainingOffset;
-            remainingOffset = 0;
-            var availableCount = layerMatched - layerOffset;
-            var layerLimit = (int)Math.Min(remainingCount, Math.Min(availableCount, int.MaxValue));
-
-            if (layerLimit <= 0)
-            {
-                continue;
-            }
-
-            plans.Add(new LayerValuePlan(
-                featureType,
-                query with { Offset = layerOffset, Limit = layerLimit },
-                layerMatched,
-                resolvedValueReference));
-
-            remainingCount -= layerLimit;
-        }
-
-        return new LayerValuePlanSet(plans.ToImmutable(), totalMatched);
-    }
 
     private async ValueTask<FeatureQuery> BuildFeatureQueryAsync(
         LayerDefinition layer,
@@ -913,56 +215,6 @@ internal sealed partial class Wfs20Handler
                 SpatialFilter = spatialFilter,
                 OutputCrs = QueryCrs.Create(outputSrid, outputAxisOrder),
                 OrderBy = orderBy
-            },
-            layer,
-            cancellationToken).ConfigureAwait(false);
-        if (!queryAdapterResult.IsSuccess || queryAdapterResult.Query == null)
-        {
-            throw new ArgumentException(queryAdapterResult.ErrorMessage ?? "Invalid WFS query parameters.");
-        }
-
-        var unifiedQuery = _queryProcessor.OptimizeQuery(queryAdapterResult.Query.Value, layer);
-        var validation = _queryProcessor.ValidateQuery(unifiedQuery, layer);
-        if (!validation.IsValid)
-        {
-            throw new ArgumentException(validation.ErrorMessage ?? "Invalid WFS query parameters.");
-        }
-
-        return _queryProcessor.ToFeatureQuery(unifiedQuery, layer);
-    }
-
-    private async ValueTask<FeatureQuery> BuildValueQueryAsync(
-        LayerDefinition layer,
-        ValueReferenceResolution valueReference,
-        string? bbox,
-        string? filter,
-        string? resourceId,
-        string? srsName,
-        bool enforceResourceIdTypeMatch,
-        bool requireResourceIdQualifier,
-        CancellationToken cancellationToken)
-    {
-        ImmutableArray<string>? outFields = valueReference.IsGeometry || valueReference.IsFeatureId
-            ? ImmutableArray<string>.Empty
-            : ImmutableArray.Create(valueReference.CanonicalName);
-
-        var (normalizedFilter, normalizedResourceId) = NormalizeFilterInputs(filter, resourceId);
-        var sqlFilter = TranslateFesFilter(layer, normalizedFilter);
-        var spatialFilter = ParseBboxFilter(bbox, layer);
-        var resourceIds = ParseResourceIds(normalizedResourceId, layer, enforceResourceIdTypeMatch, requireResourceIdQualifier);
-        sqlFilter = resourceIds.MatchesNothing
-            ? CombineSqlFilters(sqlFilter, FalseSqlFilter)
-            : sqlFilter;
-        var outputSrid = await ResolveRequestedOutputSridAsync(layer, srsName, cancellationToken).ConfigureAwait(false);
-        var outputAxisOrder = await ResolveOutputAxisOrderAsync(srsName, outputSrid, cancellationToken).ConfigureAwait(false);
-        var queryAdapterResult = await _queryParameterAdapter.ConvertAsync(
-            new Wfs20QueryRequest
-            {
-                SqlFilter = sqlFilter,
-                ObjectIds = resourceIds.ObjectIds,
-                OutFields = outFields,
-                SpatialFilter = spatialFilter,
-                OutputCrs = QueryCrs.Create(outputSrid, outputAxisOrder)
             },
             layer,
             cancellationToken).ConfigureAwait(false);
@@ -1881,117 +1133,6 @@ internal sealed partial class Wfs20Handler
         return (Results.Content(csv.ToString(), MediaTypes.Csv, Encoding.UTF8), rows.Count);
     }
 
-    private async Task<(IResult Result, int ReturnedCount)> BuildValueCollectionResultAsync(
-        LayerValuePlanSet planSet,
-        CancellationToken cancellationToken)
-    {
-        var queryResults = new List<(LayerValuePlan Plan, ImmutableArray<GmlFeature>? GmlFeatures, ImmutableArray<Feature>? Features)>(planSet.Plans.Length);
-        var returnedCount = 0;
-
-        foreach (var plan in planSet.Plans)
-        {
-            if (plan.ValueReference.IsGeometry)
-            {
-                var result = await _gmlFeatureStore.QueryGmlAsync(plan.Descriptor.Layer.Id, plan.Query, cancellationToken);
-                queryResults.Add((plan, result.Items, null));
-                returnedCount += result.Items.Length;
-                continue;
-            }
-
-            var features = await _featureReader.QueryAsync(plan.Descriptor.Layer.Id, plan.Query, cancellationToken);
-            queryResults.Add((plan, null, features.Items));
-            returnedCount += features.Items.Length;
-        }
-
-        var xml = WriteXmlDocument(writer =>
-        {
-            writer.WriteStartDocument();
-            writer.WriteStartElement("wfs", "ValueCollection", Wfs20Utilities.WfsNamespace);
-            writer.WriteAttributeString("xmlns", "gml", null, Wfs20Utilities.GmlNamespace);
-            writer.WriteAttributeString("timeStamp", DateTimeOffset.UtcNow.ToString("O", CultureInfo.InvariantCulture));
-            writer.WriteAttributeString("numberMatched", planSet.TotalMatched.ToString(CultureInfo.InvariantCulture));
-            writer.WriteAttributeString("numberReturned", returnedCount.ToString(CultureInfo.InvariantCulture));
-
-            foreach (var queryResult in queryResults)
-            {
-                if (queryResult.Plan.ValueReference.IsGeometry)
-                {
-                    foreach (var feature in queryResult.GmlFeatures ?? ImmutableArray<GmlFeature>.Empty)
-                    {
-                        writer.WriteStartElement("wfs", "member", Wfs20Utilities.WfsNamespace);
-                        if (!string.IsNullOrWhiteSpace(feature.GeometryGml))
-                        {
-                            writer.WriteRaw(feature.GeometryGml);
-                        }
-                        writer.WriteEndElement();
-                    }
-
-                    continue;
-                }
-
-                foreach (var feature in queryResult.Features ?? ImmutableArray<Feature>.Empty)
-                {
-                    writer.WriteStartElement("wfs", "member", Wfs20Utilities.WfsNamespace);
-                    var value = ExtractValue(feature, queryResult.Plan.ValueReference);
-                    if (value is not null)
-                    {
-                        writer.WriteString(ConvertValueReferenceToInvariantString(value, queryResult.Plan.ValueReference));
-                    }
-                    writer.WriteEndElement();
-                }
-            }
-            writer.WriteEndElement();
-            writer.WriteEndDocument();
-        });
-
-        return (Results.Content(xml, MediaTypes.Gml, Encoding.UTF8), returnedCount);
-    }
-
-    private async Task<(IResult Result, int ReturnedCount)> BuildValueJsonResultAsync(
-        LayerValuePlanSet planSet,
-        string normalizedFormat,
-        CancellationToken cancellationToken)
-    {
-        var features = new List<GeoJsonFeature>();
-
-        foreach (var plan in planSet.Plans)
-        {
-            var axisOrder = plan.Query.OutputAxisOrder ?? AxisOrder.EastNorth;
-            var featureResult = await _featureReader.QueryAsync(plan.Descriptor.Layer.Id, plan.Query, cancellationToken);
-            foreach (var feature in featureResult.Items)
-            {
-                var featureId = BuildFeatureId(plan.Descriptor, feature.Id);
-                var properties = new Dictionary<string, object?>(StringComparer.OrdinalIgnoreCase);
-                SimpleGeoJsonGeometry? geometry = null;
-
-                if (plan.ValueReference.IsGeometry)
-                {
-                    geometry = feature.Geometry is null
-                        ? null
-                        : _geometryServices.ConvertWkbToSimpleGeometry(feature.Geometry, axisOrder);
-                }
-                else if (plan.ValueReference.IsFeatureId)
-                {
-                    properties["id"] = featureId;
-                }
-                else
-                {
-                    properties[plan.ValueReference.CanonicalName] = ExtractValue(feature, plan.ValueReference);
-                }
-
-                features.Add(OgcGeoJsonFeatureBuilder.Create(featureId, properties, geometry));
-            }
-        }
-
-        var payload = OgcGeoJsonFeatureBuilder.CreateCollection(features, planSet.TotalMatched);
-
-        var contentType = string.Equals(normalizedFormat, Wfs20Utilities.OutputFormats.Json, StringComparison.OrdinalIgnoreCase)
-            ? MediaTypes.Json
-            : MediaTypes.GeoJson;
-
-        return (Results.Json(payload, OgcJsonContext.Default.FeatureCollection, contentType: contentType), features.Count);
-    }
-
     private static void WriteFeature(XmlWriter writer, LayerQueryPlan plan, GmlFeature feature, bool includeMemberWrapper)
     {
         if (includeMemberWrapper)
@@ -2186,32 +1327,6 @@ internal sealed partial class Wfs20Handler
             outFields.Any(candidate => candidate.Equals(field.Name, StringComparison.OrdinalIgnoreCase)));
     }
 
-    private static ValueReferenceResolution ResolveValueReference(LayerDefinition layer, string valueReference)
-    {
-        var resolvedName = FilterExpressionHelpers.ResolveFieldName(layer, valueReference, allowGeometryAlias: true)
-            ?? throw new ArgumentException($"Unknown valueReference '{valueReference}' for feature type '{layer.Name}'.");
-
-        var isGeometry = layer.GeometryField?.Name.Equals(resolvedName, StringComparison.OrdinalIgnoreCase) == true;
-        var isFeatureId = layer.PrimaryKeyField?.Name.Equals(resolvedName, StringComparison.OrdinalIgnoreCase) == true ||
-                          resolvedName.Equals("objectid", StringComparison.OrdinalIgnoreCase);
-        var field = layer.AttributeFields.FirstOrDefault(candidate =>
-            candidate.Name.Equals(resolvedName, StringComparison.OrdinalIgnoreCase));
-
-        return new ValueReferenceResolution(valueReference, resolvedName, isGeometry, isFeatureId, field);
-    }
-
-    private static object? ExtractValue(Feature feature, ValueReferenceResolution valueReference)
-    {
-        if (valueReference.IsFeatureId)
-        {
-            return feature.Id;
-        }
-
-        return feature.Attributes.TryGetValue(valueReference.CanonicalName, out var value)
-            ? value
-            : null;
-    }
-
     private static bool IsSupportedFeatureOutputFormat(string format)
     {
         return string.Equals(format, Wfs20Utilities.OutputFormats.Gml32, StringComparison.OrdinalIgnoreCase) ||
@@ -2219,14 +1334,6 @@ internal sealed partial class Wfs20Handler
                string.Equals(format, Wfs20Utilities.OutputFormats.GeoJson, StringComparison.OrdinalIgnoreCase) ||
                string.Equals(format, Wfs20Utilities.OutputFormats.Json, StringComparison.OrdinalIgnoreCase) ||
                string.Equals(format, Wfs20Utilities.OutputFormats.Csv, StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool IsSupportedValueOutputFormat(string format)
-    {
-        return string.Equals(format, Wfs20Utilities.OutputFormats.Gml32, StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(format, Wfs20Utilities.OutputFormats.Gml32Simplified, StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(format, Wfs20Utilities.OutputFormats.GeoJson, StringComparison.OrdinalIgnoreCase) ||
-               string.Equals(format, Wfs20Utilities.OutputFormats.Json, StringComparison.OrdinalIgnoreCase);
     }
 
     private static IResult CreateEmptyFeatureCollectionResult(
@@ -2300,33 +1407,6 @@ internal sealed partial class Wfs20Handler
             writer.WriteEndElement();
             writer.WriteEndDocument();
         });
-
-        return Results.Content(xml, MediaTypes.Gml, Encoding.UTF8);
-    }
-
-    private static IResult CreateEmptyValueCollectionResult(string normalizedFormat)
-    {
-        if (string.Equals(normalizedFormat, Wfs20Utilities.OutputFormats.GeoJson, StringComparison.OrdinalIgnoreCase) ||
-            string.Equals(normalizedFormat, Wfs20Utilities.OutputFormats.Json, StringComparison.OrdinalIgnoreCase))
-        {
-            var payload = OgcGeoJsonFeatureBuilder.CreateCollection(Array.Empty<GeoJsonFeature>(), 0);
-
-            var contentType = string.Equals(normalizedFormat, Wfs20Utilities.OutputFormats.Json, StringComparison.OrdinalIgnoreCase)
-                ? MediaTypes.Json
-                : MediaTypes.GeoJson;
-
-            return Results.Json(payload, OgcJsonContext.Default.FeatureCollection, contentType: contentType);
-        }
-
-        return CreateEmptyValueCollectionXmlResult();
-    }
-
-    private static IResult CreateEmptyValueCollectionXmlResult()
-    {
-        var xml = $$"""
-            <?xml version="1.0" encoding="UTF-8"?>
-            <wfs:ValueCollection xmlns:wfs="{{Wfs20Utilities.WfsNamespace}}" xmlns:gml="{{Wfs20Utilities.GmlNamespace}}" timeStamp="{{DateTimeOffset.UtcNow:O}}" numberMatched="0" numberReturned="0" />
-            """;
 
         return Results.Content(xml, MediaTypes.Gml, Encoding.UTF8);
     }
@@ -2476,11 +1556,6 @@ internal sealed partial class Wfs20Handler
             _ => ConvertStructuredValueToJson(value)
         };
     }
-
-    private static string ConvertValueReferenceToInvariantString(object? value, ValueReferenceResolution valueReference)
-        => valueReference.Field is { } field
-            ? ConvertFieldValueToInvariantString(value, field)
-            : ConvertToInvariantString(value);
 
     private static string ConvertFieldValueToInvariantString(object? value, FieldDefinition field)
     {
@@ -2701,18 +1776,8 @@ internal sealed partial class Wfs20Handler
         FeatureQuery Query,
         long MatchedCount);
 
-    private sealed record LayerValuePlan(
-        WfsFeatureTypeDescriptor Descriptor,
-        FeatureQuery Query,
-        long MatchedCount,
-        ValueReferenceResolution ValueReference);
-
     private readonly record struct LayerQueryPlanSet(
         ImmutableArray<LayerQueryPlan> Plans,
-        long TotalMatched);
-
-    private readonly record struct LayerValuePlanSet(
-        ImmutableArray<LayerValuePlan> Plans,
         long TotalMatched);
 
     private readonly record struct ResourceIdResolution(
@@ -2731,13 +1796,6 @@ internal sealed partial class Wfs20Handler
         IResult Result,
         int ReturnedCount,
         string NumberMatchedSummary);
-
-    private readonly record struct ValueReferenceResolution(
-        string RequestedName,
-        string CanonicalName,
-        bool IsGeometry,
-        bool IsFeatureId,
-        FieldDefinition? Field);
 
     private sealed class WfsQueryException(
         string exceptionCode,
